cursorAI-module7-gperez-corrected/ecommerce_api/requirements.txt
Flask==3.0.3
celery[redis]==5.4.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
redis==5.0.7
flasgger==0.9.7.1
marshmallow==3.21.3
bleach==6.1.0
python-dotenv==1.0.1
email-validator==2.1.1
fakeredis==2.23.2
pytest==8.2.2
pytest-flask==1.3.0
pytest-cov==5.0.0
